+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/ProjectChangeListener.java
@@ -1,7 +1,0 @@
-package org.jetbrains.bsp.bazel.server.sync;
-
-import org.jetbrains.bsp.bazel.server.sync.model.Project;
-
-interface ProjectChangeListener {
-  void onProjectChange(Project project);
-}